+++ CHANGELOG.rst
@@ -45,6 +45,10 @@ Changed
 Added
 -----
 
+- Added two new automated checks: SKD007 (MDI feature importance bias with
+  high-cardinality features, tip) and SKD008 (highly correlated input features,
+  issue). See :pr:`` by :user:`GaetandeCast`.
+
 Removed
 -------